Split GenericCollectionAssertionOfStringSpecs
